refactor(api): simplify feedback recording flow
